Update API Blueprint specific syntax.
Adds support for the following elements:

* Resource group `## Group <NAME>`
* Resource `## <NAME> [<URL>]`
* Resource actions `### <NAME> [<VERB>]`
* Blueprint shorthand `# <VERB> [<URL>]`
* Request/response code and type

Request/response bodies and headers are not yet highlighted in the correct
language. This may require a good bit more work to not match the `+ Body` part
while still matching the contents of the body and supporting the shorthand
where there is no explicit body list item.